transactionServer/requirements.txt
pyOpenSSL
flask
requests
